[proliferate] move cowork and mobility domains
