+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -4,15 +4,17 @@ use std::sync::Arc;
 use crate::acp::manager::AcpManager;
 use crate::adapters::git::WorkspaceFileSearchCache;
 use crate::adapters::processes::ProcessService;
-use crate::cowork::artifacts::CoworkArtifactRuntime;
-use crate::cowork::delegation::service::CoworkDelegationService;
-use crate::cowork::mcp_auth::CoworkMcpAuth;
-use crate::cowork::runtime::{CoworkRuntime, CoworkSessionHooks};
-use crate::cowork::service::CoworkService;
-use crate::cowork::store::CoworkStore;
 use crate::domains::agents::catalog::{LaunchCatalogService, ModelCatalogService};
 use crate::domains::agents::reconcile_execution::AgentReconcileService;
 use crate::domains::agents::seed::AgentSeedStore;
+use crate::domains::cowork::artifacts::CoworkArtifactRuntime;
+use crate::domains::cowork::delegation::service::CoworkDelegationService;
+use crate::domains::cowork::mcp_auth::CoworkMcpAuth;
+use crate::domains::cowork::runtime::{CoworkRuntime, CoworkSessionHooks};
+use crate::domains::cowork::service::CoworkService;
+use crate::domains::cowork::store::CoworkStore;
+use crate::domains::mobility::service::MobilityService;
+use crate::domains::mobility::store::MobilityStore;
 use crate::domains::plans::runtime::PlanRuntime;
 use crate::domains::plans::service::PlanService;
 use crate::domains::plans::store::PlanStore;
@@ -21,8 +23,6 @@ use crate::domains::reviews::mcp_auth::ReviewMcpAuth;
 use crate::domains::reviews::runtime::ReviewRuntime;
 use crate::domains::reviews::service::ReviewService;
 use crate::domains::reviews::store::ReviewStore;
-use crate::mobility::service::MobilityService;
-use crate::mobility::store::MobilityStore;
 use crate::persistence::Db;
 use crate::repo_roots::service::RepoRootService;
 use crate::repo_roots::store::RepoRootStore;
