+++ anyharness/crates/anyharness-lib/src/domains/cowork/artifacts.rs
