+++ anyharness/crates/anyharness-lib/src/domains/cowork/manifest.rs
