+++ anyharness/crates/anyharness-lib/src/domains/cowork/mcp.rs
