+++ anyharness/crates/anyharness-lib/src/domains/cowork/mcp_auth.rs
