+++ anyharness/crates/anyharness-lib/src/domains/cowork/model.rs
