+++ anyharness/crates/anyharness-lib/src/domains/mod.rs
@@ -1,3 +1,5 @@
 pub mod agents;
+pub mod cowork;
+pub mod mobility;
 pub mod plans;
 pub mod reviews;
