+++ anyharness/crates/anyharness-lib/src/lib.rs
@@ -2,10 +2,8 @@ pub mod acp;
 pub mod adapters;
 pub mod api;
 pub mod app;
-pub mod cowork;
 pub mod domains;
 pub mod integrations;
-pub mod mobility;
 pub mod origin;
 pub mod persistence;
 pub mod repo_roots;
